fixed duplicate hp bug and added random amounts of random options
Also updated the engine dataset since it was too small and therefore caused runtime exceptions

+++ src/main/java/data/engines/types/CombustionEngine.java
@@ -13,7 +13,7 @@ public class CombustionEngine extends Engine {
     private double kmPerLiter;
 
     public CombustionEngine(
-            int engineCode,
+            long engineCode,
             int hp,
             int torque,
             String fuel,
